pyproject.toml
[tool.poetry]
name = "jarxiv"
version = "0.1.0"
description = "JarXiv (Jarvis - ArXiv) is a Telegram bot designed to fetch and deliver the latest research papers from arXiv directly to users. JarXiv simplifies staying update with paprs offering daily notifications."
authors = ["Marco Gobbo <gobbo.marco98@gmail.com>"]
license = "Apache License 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^21.6"
arxiv = "^2.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
